Added modified TEM to study legacy fertilizer effects
Added xtem606cNoFert2000.cxx which assumes nitrogen fertilizers are added to croplands between 1950 and 2000, but no fertilizer is added after year 2000.  Comparison of estimates from this TEM version to those of the base TEM version allows exploration of legacy effects of nitrogen fertilizer applied to croplands during the 21st century on terrestrial biogeochemistry of abandoned croplands.